Accept clicks on workspaces borders
When the pager is pinned against screen border(s), it is very annoying
when a mouse click is ignored because it occured on the first boundary
pixel line.

Now, all workspaces borders are clickable, the bottom and right ones
too. In these cases, a click on the workspace respectively above or at
left is "emulated".